dwindow: multiple type conversion stricken. ffdshow video added. video pin and audio pin rendering normalized.
git-svn-id: svn://127.0.0.1@357 2ece6d47-a862-c445-b01b-7049066ce7d0

+++ dwindow/srt/srt_renderer.cpp
@@ -126,7 +126,7 @@ HRESULT CsrtRenderer::render(const wchar_t *text, rendered_subtitle *out, bool h
 
 	HFONT hOldFont = (HFONT) SelectObject(hdcBmp, m_font);
 
-	RECT rect = {0,0,1920,1920/m_aspect};
+	RECT rect = {0,0,1920,int(1920/m_aspect)};
 	DrawTextW(hdcBmp, text, (int)wcslen(text), &rect, DT_CENTER | DT_CALCRECT | DT_WORDBREAK | DT_NOFULLWIDTHCHARBREAK | DT_EDITCONTROL);
 	out->pixel_type = out->pixel_type_RGB;
 	out->height_pixel = rect.bottom - rect.top;
